requirements.txt
numpy==1.18.1
transformers==2.5.1
pytorch-transformers==1.2.0